+++ .github/workflows/deploy-docs.yml
@@ -136,6 +136,7 @@ jobs:
             git commit -m "Deploy preview for PR #${{ github.event.pull_request.number }}" || echo "No changes to commit"
           fi
           
+          git pull --rebase origin gh-pages
           git push origin gh-pages
 
       # Comment on PR with preview link
